Fix bug where prompt would disappear if starting new game in idle state
When removing a message from the screen, a CSS fade-out transition is played, and a timeout is set to remove the text from the DOM after the tranistion is finished. This has the un-intended side effect of deleting a new message if it is written before the timeout is triggered.

This was considered before, and I decided to just play the CSS transition and keep the text in the DOM. It has the same effect to the user, and it avoids any side-effects. However, I forgot to remove one of the timeouts and it was causing this bug.